+++ deltatau_audit/wrappers/__init__.py
@@ -1,5 +1,5 @@
 from .speed import FixedSpeedWrapper, JitterWrapper, PiecewiseSwitchWrapper
-from .latency import ObservationDelayWrapper, ActionRepeatWrapper
+from .latency import ObservationDelayWrapper, ActionRepeatWrapper, ObsNoiseWrapper
 
 __all__ = [
     "FixedSpeedWrapper",
@@ -7,4 +7,5 @@ __all__ = [
     "PiecewiseSwitchWrapper",
     "ObservationDelayWrapper",
     "ActionRepeatWrapper",
+    "ObsNoiseWrapper",
 ]